[HyperTraining] fix hypertraining and add edge cases
hypertrains properly now if pkm ivs do not match set ivs. Hypertraining is now based on the set ivs (there is a difference in cases of gen 3 and gen 4 origin pokemon which may not always have the same ivs as the set) In that case, give priority to the set. Also added edge cases for stakataka, kartana and pyukumuku. Thanks to Anubis for helping me with the level 50 edge cases and Lord Emvee (YT) for helping me with the pyukumuku edge case