add shiny paramater for alm
use Shiny: Square or Shiny: Star to specify what kind of shiny you want to gen. This filter is only applied to pokemon whose origin is gen 6 or higher. This is not applied to fixed shiny cases (ref: simpleedits.cs line 85). Also this is experimental. Report any issues on discord/gh issues