add bulk testing
create a folder called tests in the same directory. Inside the directory you can have multiple text files with showdownsets with the name in the format 'pkX - <description>.txt'. This will test the sets in the file in generation X and all future generations. All the files in the tests folder will be tested