add specific game priority
setting value of prioritizegame to false will disable priority and ignore the prioritizegameversion setting. If prioritizegame is set to true and prioritizegameversion is set to Any, priority will be given to the currently loaded save. Else, you can prioritize any game or any gamegroup from the prioritizegameversion list in settings